pyproject.toml
[tool.poetry]
name = "bookkeeper"
version = "0.1.0"
description = ""
authors = ["Daniel Grießhaber <dangrie158@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
django-stubs-ext = "^0.5.0"
django-widget-tweaks = "^1.4.12"
python-dotenv = "^0.20.0"
gunicorn = "^20.1.0"
plotly = "^5.9.0"
PyPDF2 = "^2.10.3"
pycryptodome = "^3.15.0"
aiosmtpd = {git = "https://github.com/aio-libs/aiosmtpd"}
dkimpy = "^1.0.5"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.5.0"
django-extensions = "^3.2.0"
mypy = "^0.961"
django-stubs = {extras = ["compatible-mypy"], version = "^1.12.0"}
black = {version = "^22.6.0", allow-prereleases = true}
django-browser-reload = "^1.6.0"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "bookkeeper.settings"

[tool.isort]
profile = "black"

[tool.black]
line-length = 120

[[tool.mypy.overrides]]
module = "plotly.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "dkim.*"
ignore_missing_imports = true
